+++ plugins/core/IO/qAdditionalIO/qAdditionalIO.h
@@ -18,13 +18,13 @@
 //#                                                                        #
 //##########################################################################
 
-#include "ccIOFilterPluginInterface.h"
+#include "ccIOPluginInterface.h"
 
 //! Additional I/O Formats
-class qAdditionalIO : public QObject, public ccIOFilterPluginInterface
+class qAdditionalIO : public QObject, public ccIOPluginInterface
 {
 	Q_OBJECT
-	Q_INTERFACES(ccIOFilterPluginInterface)
+	Q_INTERFACES(ccIOPluginInterface)
 	
 	Q_PLUGIN_METADATA(IID "cccorp.cloudcompare.plugin.qAdditionalIO" FILE "info.json")
 
@@ -35,7 +35,7 @@ public:
 
 	void registerCommands(ccCommandLineInterface* cmd) override;
 	
-	// inherited from ccIOFilterPluginInterface
+	// inherited from ccIOPluginInterface
 	FilterList getFilters() override;
 };