+++ plugins/core/IO/qE57IO/qE57IO.h
@@ -18,12 +18,12 @@
 //#                                                                        #
 //##########################################################################
 
-#include "ccIOFilterPluginInterface.h"
+#include "ccIOPluginInterface.h"
 
-class qE57IO : public QObject, public ccIOFilterPluginInterface
+class qE57IO : public QObject, public ccIOPluginInterface
 {
 	Q_OBJECT
-	Q_INTERFACES( ccIOFilterPluginInterface )
+	Q_INTERFACES( ccIOPluginInterface )
 	
 	Q_PLUGIN_METADATA( IID "cccorp.cloudcompare.plugin.qE57IO" FILE "info.json" )