+++ plugins/core/IO/qPhotoscanIO/qPhotoscanIO.h
@@ -18,13 +18,13 @@
 #ifndef Q_PHOTOSCAN_IO_PLUGIN_HEADER
 #define Q_PHOTOSCAN_IO_PLUGIN_HEADER
 
-#include "ccIOFilterPluginInterface.h"
+#include "ccIOPluginInterface.h"
 
 //! PhotoScan
-class qPhotoscanIO : public QObject, public ccIOFilterPluginInterface
+class qPhotoscanIO : public QObject, public ccIOPluginInterface
 {
 	Q_OBJECT
-	Q_INTERFACES(ccIOFilterPluginInterface)
+	Q_INTERFACES( ccIOPluginInterface )
 	Q_PLUGIN_METADATA(IID "cccorp.cloudcompare.plugin.qPhotoscanIO" FILE "info.json")
 
 public:
@@ -32,7 +32,7 @@ public:
 	
 	~qPhotoscanIO() override = default;
 
-	//inherited from ccIOFilterPluginInterface
+	//inherited from ccIOPluginInterface
 	FilterList getFilters() override;
 };